Fixing my stupidity and adding some mounts

+++ src/NosCore.GameObject/Providers/GuriProvider/Handlers/EmoticonHandler.cs
@@ -38,7 +38,7 @@ namespace NosCore.GameObject.Providers.GuriProvider.Handlers
             if (requestData.Data.VisualEntityId.GetValueOrDefault() == requestData.ClientSession.Character.CharacterId)
             {
                 requestData.ClientSession.Character.MapInstance.Sessions.SendPacket(
-                    requestData.ClientSession.Character.GenerateEff(requestData.Data.Data +
+                    requestData.ClientSession.Character.GenerateEff((int)requestData.Data.Data +
                         4099)); //TODO , ReceiverType.AllNoEmoBlocked
             }
         }
